+++ src/lib/components/ExplorerLink.tsx
@@ -4,8 +4,8 @@ import { Box, Text } from "@chakra-ui/react";
 import type { ExplorerConfig } from "config/types";
 import type { AddressReturnType } from "lib/app-provider";
 import { useCelatoneApp } from "lib/app-provider/contexts";
+import { useBaseApiRoute } from "lib/app-provider/hooks/useBaseApiRoute";
 import { useCurrentChain } from "lib/app-provider/hooks/useCurrentChain";
-import { useLCDEndpoint } from "lib/app-provider/hooks/useLCDEndpoint";
 import { AmpTrackMintscan } from "lib/services/amplitude";
 import type { Option } from "lib/types";
 import { truncate } from "lib/utils";
@@ -18,7 +18,8 @@ export type LinkType =
   | "tx_hash"
   | "code_id"
   | "block_height"
-  | "proposal_id";
+  | "proposal_id"
+  | "pool_id";
 
 interface ExplorerLinkProps extends BoxProps {
   value: string;
@@ -66,6 +67,9 @@ export const getNavigationUrl = (
         explorerConfig.proposal ||
         `${lcdEndpoint}/cosmos/gov/v1beta1/proposals`;
       break;
+    case "pool_id":
+      url = "/pools";
+      break;
     case "invalid_address":
       return "";
     default:
@@ -88,7 +92,7 @@ const getValueText = (
 const getCopyLabel = (type: LinkType) =>
   type
     .split("_")
-    .map((str) => str.charAt(0).toUpperCase() + str.slice(1))
+    .map((str: string) => str.charAt(0).toUpperCase() + str.slice(1))
     .join(" ");
 
 const LinkRender = ({
@@ -159,7 +163,7 @@ export const ExplorerLink = ({
   ...componentProps
 }: ExplorerLinkProps) => {
   const { address } = useCurrentChain();
-  const lcdEndpoint = useLCDEndpoint();
+  const lcdEndpoint = useBaseApiRoute("rest");
   const {
     chainConfig: { explorerLink: explorerConfig },
   } = useCelatoneApp();
@@ -169,7 +173,8 @@ export const ExplorerLink = ({
     type === "contract_address" ||
     type === "user_address" ||
     type === "tx_hash" ||
-    type === "block_height";
+    type === "block_height" ||
+    type === "pool_id";
 
   const [hrefLink, textValue] = [
     getNavigationUrl(type, explorerConfig, copyValue || value, lcdEndpoint),
@@ -193,7 +198,9 @@ export const ExplorerLink = ({
       {...componentProps}
     >
       {readOnly ? (
-        <Text variant="body2">{textValue}</Text>
+        <Text variant="body2" color="text.disabled">
+          {textValue}
+        </Text>
       ) : (
         <>
           <LinkRender